refactor dbt rpc in prep for async
Split rpc.py -> rpc/...
Move async entries in node_runners.py -> rpc/node_runners.py
Move RemoteCallable -> rpc/task.py
Move all RPC tasks -> task/remote.py
Clean up results for kill into a more coherent hologram type structure
Increased some test timeouts to improve some flaky tests around start-up compilation
Add many more type annotations for mypy
