+++ core/dbt/rpc/__init__.py
@@ -1,5 +1,6 @@
-from dbt.rpc.task_manager import TaskManager  # noqa
-from dbt.rpc.response_manager import ResponseManager  # noqa
 from dbt.rpc.error import (  # noqa
     dbt_error, server_error, invalid_params, RPCException
 )
+from dbt.rpc.task import RemoteCallable, RemoteCallableResult # noqa
+from dbt.rpc.task_manager import TaskManager  # noqa
+from dbt.rpc.response_manager import ResponseManager  # noqa
