+++ core/dbt/task/runnable.py
@@ -1,12 +1,8 @@
-import base64
 import os
 import time
-from abc import abstractmethod
 from datetime import datetime
 from multiprocessing.dummy import Pool as ThreadPool
-from typing import Union, List, Optional
 
-from dbt import rpc
 from dbt.task.base import ConfiguredTask
 from dbt.adapters.factory import get_adapter
 from dbt.logger import GLOBAL_LOGGER as logger
@@ -337,51 +333,3 @@ class GraphRunnableTask(ManifestTask):
 
     def task_end_messages(self, results):
         dbt.ui.printer.print_run_end_messages(results)
-
-
-class RemoteCallable:
-    METHOD_NAME: Optional[str] = None
-    is_async = False
-
-    @abstractmethod
-    def handle_request(self):
-        raise dbt.exceptions.NotImplementedException(
-            'from_kwargs not implemented'
-        )
-
-    @staticmethod
-    def _listify(
-        value: Optional[Union[str, List[str]]]
-    ) -> Optional[List[str]]:
-        if value is None:
-            return None
-        elif isinstance(value, str):
-            return [value]
-        else:
-            return value
-
-    def decode_sql(self, sql):
-        """Base64 decode a string. This should only be used for sql in calls.
-
-        :param str sql: The base64 encoded form of the original utf-8 string
-        :return str: The decoded utf-8 string
-        """
-        # JSON is defined as using "unicode", we'll go a step further and
-        # mandate utf-8 (though for the base64 part, it doesn't really matter!)
-        base64_sql_bytes = str(sql).encode('utf-8')
-
-        try:
-            sql_bytes = base64.b64decode(base64_sql_bytes, validate=True)
-        except ValueError:
-            self.raise_invalid_base64(sql)
-
-        return sql_bytes.decode('utf-8')
-
-    @staticmethod
-    def raise_invalid_base64(sql):
-        raise rpc.invalid_params(
-            data={
-                'message': 'invalid base64-encoded sql input',
-                'sql': str(sql),
-            }
-        )
