+++ core/dbt/task/seed.py
@@ -1,11 +1,9 @@
 import random
-from typing import List, Dict, Any
 
 from dbt.logger import GLOBAL_LOGGER as logger
 from dbt.node_runners import SeedRunner
 from dbt.node_types import NodeType
 from dbt.task.run import RunTask
-from dbt.task.runnable import RemoteCallable
 import dbt.ui.printer
 
 
@@ -47,21 +45,3 @@ class SeedTask(RunTask):
         for result in results:
             if result.error is None:
                 self.show_table(result)
-
-
-class RemoteSeedProjectTask(SeedTask, RemoteCallable):
-    METHOD_NAME = 'seed_project'
-
-    def __init__(self, args, config, manifest):
-        super().__init__(args, config)
-        self.manifest = manifest.deepcopy(config=config)
-
-    def load_manifest(self):
-        # we started out with a manifest!
-        pass
-
-    def handle_request(self, show: bool = False) -> Dict[str, List[Any]]:
-        self.args.show = show
-
-        results = self.run()
-        return {'results': [r.to_dict() for r in results]}
