+++ core/dbt/task/test.py
@@ -1,9 +1,6 @@
-from typing import Union, List, Dict, Any
-
 from dbt.node_runners import TestRunner
 from dbt.node_types import NodeType
 from dbt.task.run import RunTask
-from dbt.task.runnable import RemoteCallable
 
 
 class TestTask(RunTask):
@@ -41,30 +38,3 @@ class TestTask(RunTask):
 
     def get_runner_type(self):
         return TestRunner
-
-
-class RemoteTestProjectTask(TestTask, RemoteCallable):
-    METHOD_NAME = 'test_project'
-
-    def __init__(self, args, config, manifest):
-        super().__init__(args, config)
-        self.manifest = manifest.deepcopy(config=config)
-
-    def load_manifest(self):
-        # we started out with a manifest!
-        pass
-
-    def handle_request(
-        self,
-        models: Union[None, str, List[str]] = None,
-        exclude: Union[None, str, List[str]] = None,
-        data: bool = False,
-        schema: bool = False,
-    ) -> Dict[str, List[Any]]:
-        self.args.models = self._listify(models)
-        self.args.exclude = self._listify(exclude)
-        self.args.data = data
-        self.args.schema = schema
-
-        results = self.run()
-        return {'results': [r.to_dict() for r in results]}
